+++ templates/chat.html
@@ -36,7 +36,7 @@
             userInput.value = "";
         
             try {
-                // ✅ Flask ke backend ke liye sahi request bhejna
+                //  Flask ke backend ke liye sahi request bhejna
                 let response = await fetch("http://127.0.0.1:5000/auth/chat", {
                     method: "POST",
                     headers: { "Content-Type": "application/json" },
