+++ templates/mainpage.html
@@ -63,7 +63,7 @@
                 }
             });
 
-            // ✅ Upload File Event
+            //  Upload File Event
             $("#upload-button").click(function (event) {
                 event.preventDefault();  // reload stop
 
@@ -104,7 +104,7 @@
                 });
             });
 
-            // ✅ Summarize Button
+            //  Summarize Button
             $("#summarize").click(function (event) {
                 event.preventDefault();
